more cards yaaaa
added the more cards like spice, all purpose and everything nice to create the perfect cards. i also a set minimum and max amount of damage

+++ KKC-GAMEDEV-CLUB-2D/Assets/Scripts/card scripts/CardProjectile.cs
@@ -7,10 +7,10 @@ public class CardProjectile : MonoBehaviour
     private Vector3 targetPos;
     private float speed = 1f;
     private float t = 0f;
-    private int damage;
+    private float damage;
     private bool isLaunched = false;
 
-    public void Launch(Vector3 start, Vector3 control, Vector3 target, int cardDamage)
+    public void Launch(Vector3 start, Vector3 control, Vector3 target, float cardDamage)
     {
         startPos = start;
         controlPoint = control;
